requirements.txt
rauth==0.7.2
simplejson==3.8.1
websocket-server==0.4
django-debug-toolbar==1.4
django-excel
import pyexcel.ext.xls
nose
rednose
nose-cov
codecov
coverage
lxml
pyexcel-ods3>=0.1.0
pyexcel-xls>=0.1.0
pyexcel-xlsx>=0.1.0
